requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.9.4
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
bidict==0.22.1
blobfile==2.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
dnspython==1.16.0
eventlet==0.30.2
filelock==3.9.0
Flask==2.2.3
Flask-Executor==1.0.0
Flask-Migrate==2.6.0
Flask-Script==2.0.6
Flask-SocketIO==5.3.3
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.0
itchat==1.3.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
Mako==1.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
multidict==6.0.4
nltk==3.8.1
openai==0.27.0
pip-autoremove==0.10.0
pycryptodomex==3.17
pypng==0.20220715.0
PyQRCode==1.2.1
PySocks==1.7.1
python-engineio==4.3.4
python-socketio==5.7.2
python-telegram-bot==20.1
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.4
tiktoken==0.3.0
tqdm==4.64.1
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
